+++ src/Revit.TestRunner/Main.cs
@@ -5,21 +5,27 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using Autodesk.Revit.UI;
 using Revit.TestRunner.Commands;
+using Revit.TestRunner.Server;
 
 namespace Revit.TestRunner
 {
+    /// <summary>
+    /// Entreance object for the Revit.TestRunner addin in Revit.
+    /// </summary>
     public class Main : IExternalApplication
     {
+        private RunnerController mController;
+
         public Result OnStartup( UIControlledApplication application )
         {
-            Log.Info( $"Revit.TestRunner started '{DateTime.Now}'" );
+            Log.Info( $"Revit.TestRunner started v{Assembly.GetExecutingAssembly().GetName().Version} '{DateTime.Now}'" );
             Log.Info( $"{Environment.OSVersion}, NetFX {Environment.Version}" );
             Log.Debug( $"Log Directory '{Log.LogDirectory}'" );
             Log.Debug( $"CurrentAppDomain.ApplicationBase '{AppDomain.CurrentDomain.SetupInformation.ApplicationBase}'" );
 
             RibbonPanel ribbonPanel = application.CreateRibbonPanel( "Testing" );
 
-            string command = typeof( TestRunnerCommand ).FullName;
+            string command = typeof( RunnerCommand ).FullName;
 
             PushButtonData buttonData = new PushButtonData( command, "Open Runner", Assembly.GetExecutingAssembly().Location, command ) {
                 ToolTip = "Open the Test Runner Dialog\nStart Tests using the Revit API.",
@@ -30,11 +36,15 @@ namespace Revit.TestRunner
 
             ribbonPanel.AddItem( buttonData );
 
+            mController = new RunnerController();
+            mController.Start( application );
+
             return Result.Succeeded;
         }
 
         public Result OnShutdown( UIControlledApplication application )
         {
+            mController.Stop();
             return Result.Succeeded;
         }
 
@@ -43,7 +53,7 @@ namespace Revit.TestRunner
             ImageSource result = null;
 
             try {
-                string ressource = "pack://application:,,,/Revit.TestRunner;component/View/Pic/" + name;
+                string ressource = "pack://application:,,,/Revit.TestRunner;component/Commands/Pic/" + name;
                 result = new BitmapImage( new Uri( ressource ) );
                 result = Thumbnail( result, size );
             }